// src/types/v9111.ts
export type H256 = Uint8Array
// ...
export interface ReferendumInfo_Ongoing {
  __kind: 'Ongoing'
  value: ReferendumStatus
}

export interface ReferendumInfo_Finished {
  __kind: 'Finished'
  approved: boolean
  end: number
}
// ...
export interface ReferendumStatus {
  end: number
  proposalHash: H256
  threshold: VoteThreshold
  delay: number
  tally: Tally
}
// ...
export type VoteThreshold = VoteThreshold_SuperMajorityApprove | VoteThreshold_SuperMajorityAgainst | VoteThreshold_SimpleMajority

export interface VoteThreshold_SuperMajorityApprove {
  __kind: 'SuperMajorityApprove'
}

export interface VoteThreshold_SuperMajorityAgainst {
  __kind: 'SuperMajorityAgainst'
}

export interface VoteThreshold_SimpleMajority {
  __kind: 'SimpleMajority'
}

export interface Tally {
  ayes: bigint
  nays: bigint
  turnout: bigint
}